Fix bug
Bug 1: Export Procedure & Function - If the MySQL user is different than definer, the create command is blank. MySqlBackup will export it with "DROP" but without "CREATE". Now, if the create command is empty, no "DROP" will be generated.
Bug 2: When ImportInfo.IgnoreSqlError = true and an error occur, the current error SQL statement is not cleared and resulting the following whole sql document to be error.